+++ models/dailyWords.js
@@ -1,1 +1,2 @@
-export const DAILY_LIRDLE = 'notch';
+export const DAILY_LIRDLE = 'thank';
+export const DAILY_WORDLE = 'pecan';