+++ test/dailyWordsTest.js
@@ -1,3 +1,0 @@
-import { DAILY_LIRDLE } from '../models/dailyWords.js';
-
-console.log('DAILY_LIRDLE:', DAILY_LIRDLE);